+++ BWBE/Models/Ingredient.cs
@@ -9,6 +9,6 @@ public class Ingredient
     [MaxLength(36)] public string InventoryId { get; set; } = null!;
     [MaxLength(50)] public string Name { get; set; } = null!;
     public float Quantity { get; set; }
-    public int MinQuantity { get; set; }
+    public float MinQuantity { get; set; }
     [MaxLength(50)] public string Unit { get; set; } = null!;
 }
